Changes to be committed: 	new file:   data/graphs_pubs/graph-107.net 	new file:   data/graphs_pubs/graph-131.net 	new file:   data/graphs_pubs/graph-144.net 	new file:   data/graphs_pubs/graph-145.net 	new file:   data/graphs_pubs/graph-16.net 	new file:   data/graphs_pubs/graph-162.net 	new file:   data/graphs_pubs/graph-182.net 	new file:   data/graphs_pubs/graph-199.net 	new file:   data/graphs_pubs/graph-24.net 	new file:   data/graphs_pubs/graph-75.net 	new file:   data/movies/newmovies.txt 	modified:   graph.cpp 	modified:   graph.hpp 	modified:   main.cpp 	modified:   makefile 	new file:   results/data.txt 	new file:   sumprod.cpp 	new file:   sumprod.hpp 	modified:   test.cpp 	modified:   test.hpp 	new file:   test_data/testgraph.txt 	new file:   test_data/testgraph2.txt 	new file:   test_run.cpp
Added in the data files from the paper in the data directory.

Have tested and ensured that all data in test suite and from the newmovies.txt are read in and deleted properly.
Results for timing and anything else will be printed to a file in the results dir.
Test data was moved to a test_data dir.
Now using smart pointers to ensure no memory leaks occur.
Introduced files for sumprod alg and working on implementing those now.